依存poiは結局3.1-FINALに戻しました。
git-svn-id: https://www.seasar.org/svn/sandbox/fisshplate/trunk@224 c159f9c0-7f7c-4dca-964d-717b2124a8cb

+++ fisshplate/src/test/java/org/seasar/fisshplate/core/parser/LinkParserTest.java
@@ -39,7 +39,7 @@ public class LinkParserTest extends TestCase {
 
     public void testパース成功(){
         HSSFWorkbook wb = new HSSFWorkbook();
-        wb.createSheet().createRow(0).createCell(0);
+        wb.createSheet().createRow(0).createCell((short)0);
         WorkbookWrapper ww = new WorkbookWrapper(wb);
         
         CellWrapper cell = ww.getSheetAt(0).getRow(0).getCell(0);
@@ -66,7 +66,7 @@ public class LinkParserTest extends TestCase {
     
     public void testパースはずれ(){
         HSSFWorkbook wb = new HSSFWorkbook();
-        wb.createSheet().createRow(0).createCell(0);
+        wb.createSheet().createRow(0).createCell((short)0);
         WorkbookWrapper ww = new WorkbookWrapper(wb);
         
         CellWrapper cell = ww.getSheetAt(0).getRow(0).getCell(0);
